requirements.txt
Django==3.2.6
django_filter==21.1
djangorestframework==3.13.1
